+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### Tests
+
+- Added unit tests for `fastmath.h`
+
 ## [1.4.0] - 2022-05-07
 
 ### Added